+++ HISTORY.rst
@@ -9,6 +9,9 @@ History
 - Reload local file url and other metadata after uploading to Rana.
 - Show "Latest revision number" field in 3Di schematisation file details. (nens/rana#1874)
 - Show description field.
+- Add option to pass Rana URL to QGIS. (nens/rana-qgis-plugin#61)
+- Add option to load WMS layers of 3Di scenarios (nens/rana-qgis-plugin#51)
+- Add option to download scenario zip (nens/rana-qgis-plugin#51)
 - Reload file ui after uploading file. (nens/rana#41)
 - Add refresh button to project and file overview to allow manual reloading.